Update README.md
# COMP213_Assessment1
# The Card Game
The first round of the game starts with all players having the same number of cards in their deck, and the cards are facing down (i.e. not visible to the players); all cards have a number of attributes, each attribute having a value from 0 up to 9 (inclusive).
Every player picks up a card from the top of their deck and looks at it.
The nominated first player then chooses an attribute to “play with”; all the players compare the values of this Attribute on the cards they have picked up.
The player with the highest value wins the round, collects the cards that the other players had picked up and puts them (together with his/her card) facing down at the bottom of his/her deck.
The next round starts with every player picking up the card on top of his/her deck and looking at it, and a next player (or random player or the previous winner etc) chooses the attribute they wish to play with.
The game continues in the same fashion for an unlimited number of rounds until only one player, the winner, has cards left in their deck. This game also supports multi-players (up to 5 players).